Added test for delete task

+++ backend/src/actions/task/DeleteTaskAbstract.js
@@ -11,27 +11,41 @@ class DeleteTaskAbstract extends ActionBase {
 
         this.#validate();
 
-        const transaction = sequelize.transaction({ autocommit: false });
+        const transaction = await sequelize.transaction({ autocommit: false });
 
         try {
+            const task = await Task.findOne({ where: { id: this.taskId } })
+
+            if (!task) {
+                throw new this.app.TransportError({
+                    message: 'Not found task for delete',
+                    status: 404,
+                });
+            }
+
             await Task.destroy({
                 where: { id: this.taskId },
                 transaction,
             })
 
             await transaction.commit();
+
+            return {
+                message: 'Task deleted',
+            };
         } catch (error) {
             await transaction.rollback();
 
             const message = 'Error on delete task';
             throw new this.app.TransportError({
-                message,
-                status: 400,
-                error,
+                message: error.message || message,
+                status: error.status || 400,
+                ...!error.message && { error },
             });
         }
     }
-    // check some dependencies before delete task
+
+    // check dependencies before delete task
     #validate () {
         return true;
     }
